requirements.txt
click==8.1.7
clearml==1.12.2
loguru==0.7.2
matplotlib==3.8.2
numpy==1.25.0
opencv-python==4.8.1.78
pandas==2.1.4
plotly==5.22.0
pydantic==2.1.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytorch-lightning[extra]==2.1.3
scikit_learn==1.5.2
scipy==1.11.4
seaborn==0.13.2
tensorboard==2.15.1
tensorboardX==2.6.2.2
torch==2.1.0
torcheval==0.0.7
torchtext==0.16.0
torchvision==0.16.0
tqdm==4.66.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
